move interaction test
