+++ packages/table/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@leafygreen-ui/table",
-  "version": "12.7.0-test.5",
+  "version": "12.7.0-test.9",
   "description": "leafyGreen UI Kit Table",
   "main": "./dist/index.js",
   "module": "./dist/esm/index.js",
@@ -46,8 +46,7 @@
     "@leafygreen-ui/button": "^21.2.0",
     "@leafygreen-ui/badge": "^8.1.2",
     "@leafygreen-ui/pagination": "^1.0.23",
-    "@lg-tools/storybook-utils": "^0.1.1",
-    "@emotion/styled": "^11.10.5"
+    "@lg-tools/storybook-utils": "^0.1.1"
   },
   "peerDependencies": {
     "@leafygreen-ui/leafygreen-provider": "^3.1.12"
